(fix Tekla Structures) PluginCallback Small fixes  (#1395)
* Tekla Beam to Revit

* Change to non dependent plugin

* Fix several issues

Fix winforms window closing, serialisation of saved streams, and exception on Tekla close.

* Small fixes. Match 2020 plugin changes to 2021

* (fix) Fixing Tekla Structures 2020 and 2021 plugin callback - winforms

* fix for merge
