+++ ConnectorTeklaStructures/ConnectorTeklaStructures2020/MainForm.cs
@@ -1,11 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.IO;
 
 using Avalonia;
@@ -127,5 +120,10 @@ namespace Speckle.ConnectorTeklaStructures
             {
             }
         }
+
+    private void MainForm_Load(object sender, EventArgs e)
+    {
+
     }
+  }
 }
