+++ Objects/Objects/BuiltElements/Beam.cs
@@ -73,10 +73,13 @@ namespace Objects.BuiltElements.TeklaStructures
     public SectionProfile profile { get; set; }
     [DetachProperty]
     public Material material { get; set; }
-    public string finish { get; set; }
-    public string classNumber { get; set; }
+        [DetachProperty]
+        public string finish { get; set; }
+        [DetachProperty]
+        public string classNumber { get; set; }
     public Vector alignmentVector { get; set; } // This can be set to get proper rotation if coming from an application that doesn't have positioning
-    public TeklaPosition position { get; set; }
+        [DetachProperty]
+        public TeklaPosition position { get; set; }
     public Base userProperties { get; set; }
 
     [DetachProperty]
